fix: fixed BME280 pressure formatting for values <1000

+++ src/bme280.rs
@@ -50,7 +50,8 @@ impl BME280Sensor {
             .map_err(|e| anyhow::anyhow!("Failed to read BME280 sensor: {:?}", e))?;
 
         let temperature = (measurements.temperature * 2.0).round() as u8;
-        let pressure = ((measurements.pressure / 100.0) - 1000.0).round() as u8;
+        // cast pressure to i8 (two's complement), then fit into u8 to send through LoRa
+        let pressure = ((measurements.pressure / 100.0) - 1000.0).round() as i8 as u8;
         let humidity = measurements.humidity.round() as u8;
     
         Ok(PacketBME280 {
@@ -62,7 +63,7 @@ impl BME280Sensor {
 
     pub fn print(&self, data: &PacketBME280) -> Result<()> {
         let temperature = data.temperature as f32 / 2.0;
-        let pressure = (data.pressure as f32 + 1000.0) * 100.0;
+        let pressure = (data.pressure as i8 as f32 + 1000.0) * 100.0;
         let humidity = data.humidity as f32;
 
         info!("BME280 Sensor Measurements:");
